% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilexicon_update.R
\name{bilexicon_update}
\alias{bilexicon_update}
\title{Retrieve and update local and/or remote data from formr}
\usage{
bilexicon_update(
  email = "gonzalo.garciadecastro@upf.edu",
  bilingual_threshold = 95,
  runs = c("formr-2", "formr-short", "formr-lockdown")
)
}
\arguments{
\item{email}{E-mail used in formr account.}

\item{bilingual_threshold}{Minimum degree of exposure to consider a participant as *monolingual*.}

\item{runs}{formr runs to update}
}
\value{
A list with 3 elements: \code{responses} contains the each participant's response to each item, along with some session-specific metadata. \code{participants} contains participant-level data, such as demographic and linguitic information, and some metadata about their session. \code{summary} summarises the demographic and linguistic characteristics of the sample.
}
\description{
Retrieve and update local and/or remote data from formr
}
\examples{
d <- bilexicon_update(email = "user@mail.com", bilingual_threshold = 95, runs = c("formr-short", "formr-2"))
}
