% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_childes_frequency}
\alias{get_childes_frequency}
\title{Extract lexical frequencies from CHILDES}
\usage{
get_childes_frequency(token, languages = c("cat", "spa"), ...)
}
\arguments{
\item{token}{Character string vector indicating the words-forms to look up}

\item{languages}{Languages in which to look up the word forms in ISO code (see ISOcodes::ISO_639_2). Default are c("cat", and "spa"), Catalan and Spanish.}

\item{...}{Additional arguments passed to the \code{childesr::get_speaker_statistics} function. Use it to refine the search.}
}
\description{
Extract lexical frequencies from CHILDES
}
