% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_formr2.R
\name{import_formr2}
\alias{import_formr2}
\title{Import formr 2.0}
\usage{
import_formr2(
  surveys = c("bilexicon_01_log", "bilexicon_02_welcome", "bilexicon_03_consent",
    "bilexicon_04_demo", "bilexicon_05_language", "bilexicon_06_words_cat",
    "bilexicon_06_words_spa")
)
}
\description{
Import formr 2.0
}
