% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_formr-lockdown.R
\name{import_formr_lockdown}
\alias{import_formr_lockdown}
\title{Import lockdown}
\usage{
import_formr_lockdown(
  surveys = c("bilexicon_lockdown_01_log", "bilexicon_lockdown_02_welcome",
    "bilexicon_lockdown_03_consent", "bilexicon_lockdown_04_demo",
    "bilexicon_lockdown_05_language", "bilexicon_lockdown_06_words_catalan",
    "bilexicon_lockdown_06_words_spanish")
)
}
\description{
Import lockdown
}
