% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_pool}
\alias{import_pool}
\title{Import pool}
\usage{
import_pool(file = system.file("extdata", "pool.xlsx", package = "multilex"))
}
\arguments{
\item{file}{Path of the pool.xlsx file}
}
\description{
Import pool
}
