% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_add_participants.R
\name{ml_add_participants}
\alias{ml_add_participants}
\title{Point-and-click interface for adding participants to \code{Participants}}
\usage{
ml_add_participants(google_email = NULL, participants = NULL)
}
\arguments{
\item{google_email}{E-mail used in Google Drive account.}

\item{participants}{Participants data frame, as generated by \code{ml_participants}. If NULL (default), \code{ml_participants} is run.}
}
\value{
A GUI that opens in the default browser
}
\description{
Point-and-click interface for adding participants to \code{Participants}
}
\examples{
ml_add_participants(email = "user@mail.com")

}
