% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_logs_summary.R
\name{ml_logs_summary}
\alias{ml_logs_summary}
\title{Summarise participant information}
\usage{
ml_logs_summary(logs = NULL)
}
\arguments{
\item{logs}{Logs, as generated by \code{ml_logs}. If NULL (default), \code{ml_logs} is run.}
}
\value{
A data frame with the sample size of each group, and the list of IDs in each group.
}
\description{
Summarise participant information
}
\examples{
d <- ml_logs_summary(responses = ml_responses)
}
