% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_norms.R
\name{ml_norms}
\alias{ml_norms}
\title{Generate item-level norms for age, sex, language profile and item dominance.}
\usage{
ml_norms(
  participants = NULL,
  responses = NULL,
  item = NULL,
  language = c("Catalan", "Spanish"),
  type = c("understands", "produces"),
  age = c(0, 100),
  lp = c("Bilingual", "Monolingual", "Other"),
  sex = c("Female", "Male"),
  category = NULL,
  .width = 0.95
)
}
\arguments{
\item{participants}{Participants data frame, as generated by
\code{ml_participants}. If NULL (default), \code{ml_participants} is run.}

\item{responses}{Responses data frame, as generated by \code{ml_responses}.
If NULL (default), \code{ml_responses} is run.}

\item{item}{Character string indicating the item to compute norms for. If
left NULL (by default) norms will be computed for all items. You can check
the available items in the \code{pool} data set running \code{data("pool")}.}

\item{language}{Character string indicating the language to compute
vocabulary norms for: "catalan" and/or
"spanish"}

\item{type}{Character string indicating the vocabulary type to compute norms
for. Takes "understands" and/or"produces" (defaults to both).}

\item{age}{Numeric vector of length two (min-max) indicating the age range of
participants to compute norms for.}

\item{lp}{character string indicating the language profile of participants to
compute norms for: "Bilingual", "Monolingual", "Other" (defaults to all).}

\item{sex}{character string indicating the sex of participants to compute
norms for. Takes "Female" and/or "Male" (defaults to both).}

\item{category}{character string indicating the semantic/functional
category(ies) to include items from. See available categories in the
\code{pool} data set by running \code{data("pool")}.}

\item{.width}{Numeric values ranging from 0 to 1 (not included) indicating
the confidence level of confidence intervals (defaults to 0.95).}
}
\value{
A data frame with the proportion of participants in the sample that
understand or produce the items indicated in \code{item}, along with the
standard error and confidence interval of the estimation.
}
\description{
Generate item-level norms for age, sex, language profile and item dominance.
}
\details{
Estimated proportions and corresponding standard errors and
confidence intervals are computed following
\insertCite{gelman2020regression;textual}{multilex}'s adjustments to account for
zero- and one-inflation (see functions \code{prop_adj}, \code{prop_adj_se}, and
\code{prop_adj_ci}).
}
\examples{
ml_norms(item = "cat_casa", type = "understands", age = c(20, 24))
}
\references{
\insertRef{gelman2020regression}{multilex}
}
