% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_participants.R
\name{ml_participants}
\alias{ml_participants}
\title{Retrieve and update local and/or remote data from formr}
\usage{
ml_participants(google_email = NULL)
}
\arguments{
\item{google_email}{E-mail used in Google Drive account.}
}
\value{
A data frame with all participants that have participated or are candidates to participate in any of the versions of MultiLex
}
\description{
Retrieve and update local and/or remote data from formr
}
\examples{
participants <- ml_participants(email = "user@mail.com")

}
