% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_responses.R
\name{ml_responses}
\alias{ml_responses}
\title{Retrieve and update local and/or remote data from formr}
\usage{
ml_responses(
  participants = NULL,
  formr_email = "gonzalo.garciadecastro@upf.edu",
  google_email = NULL,
  runs = c("BL-Long2", "BL-Lockdown")
)
}
\arguments{
\item{participants}{Participants data frame, as generated by \code{ml_participants}. If NULL (default), \code{ml_participants} is run.}

\item{runs}{formr runs to update}
}
\value{
A dataframe containing participant's responses to each item, along with some session-specific metadata.
}
\description{
Retrieve and update local and/or remote data from formr
}
\examples{
d <- ml_responses(runs = c("formr-short", "formr-2"))
}
