% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_vocabulary.R
\name{ml_vocabulary}
\alias{ml_vocabulary}
\title{Generate participant information and progress for each response}
\usage{
ml_vocabulary(
  participants = NULL,
  responses = NULL,
  google_email = NULL,
  vocab_type = c("understands", "produces"),
  vocab_language = c("Catalan", "Spanish"),
  keep_cols = NULL
)
}
\arguments{
\item{participants}{Participants data frame, as generated by \code{ml_participants}. If NULL (default), \code{ml_participants} is run.}

\item{responses}{Responses data frame, as generated by \code{ml_responses}. If NULL (default), \code{ml_responses} is run.}

\item{google_email}{E-mail used in Google Drive account.}

\item{vocab_type}{Vocabulary type: "comprehensive" or "productive"}

\item{vocab_language}{Language to compute vocabulary sizes for: "catalan" and/or "spanish}

\item{keep_cols}{Columns to be kept in the output dataframe (apart from \code{id}, \code{time} and \code{language}).}
}
\value{
A dataset with each participant's comprehensive and/or vocabulary size in each language
}
\description{
Generate participant information and progress for each response
}
\examples{
vocabulary <- ml_vocabulary(google_email = "user@email.com"))

}
