% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_adj}
\alias{prop_adj}
\title{Adjusted proportion (Gelman, Hill & Vehtari, 2020)}
\usage{
prop_adj(x, n)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of tries}
}
\description{
Adjusted proportion (Gelman, Hill & Vehtari, 2020)
}
