% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_adj_ci}
\alias{prop_adj_ci}
\title{Adjusted confidence interval of proportion (Gelman, Hill & Vehtari, 2020)}
\usage{
prop_adj_ci(x, n, .width = 0.95)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of tries}

\item{.width}{Confidence level (defaults to .95)}
}
\description{
Adjusted confidence interval of proportion (Gelman, Hill & Vehtari, 2020)
}
