% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_adj_se}
\alias{prop_adj_se}
\title{Adjusted standard error of proportion (Gelman, Hill & Vehtari, 2020)}
\usage{
prop_adj_se(x, n)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of tries}
}
\description{
Adjusted standard error of proportion (Gelman, Hill & Vehtari, 2020)
}
