% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_app.R
\name{ml_app}
\alias{ml_app}
\title{Point-and-click interface for MultiLex}
\usage{
ml_app(participants = NULL, responses = NULL)
}
\arguments{
\item{participants}{Participants data frame, as generated by \code{ml_participants}. If NULL (default), \code{ml_participants} is run.}

\item{responses}{Responses data frame, as generated by \code{ml_responses}. If NULL (default), \code{ml_responses} is run.}
}
\value{
A GUI that opens MultiLex in the default browser
}
\description{
Point-and-click interface for MultiLex
}
\examples{
ml_app()

}
