% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_connect.R
\name{ml_connect}
\alias{ml_connect}
\title{Authenticate in Google and formr}
\usage{
ml_connect(google_email = NULL)
}
\arguments{
\item{google_email}{E-mail used in Google Drive account. If NULL (default), it is assumed to be the same as admin email (gonzalo.garciadecastro@upf.edu)}
}
\value{
TRUE if Google and formr authentication was successful, FALSE if authentication of any of the two failed.
}
\description{
Authenticate in Google and formr
}
\details{
This function authenticates both the Google and formr accounts.
It takes \code{google_email}(the email address you use in your Google Drive account,
where the \code{Participants} database is stored) as a single argument.
You will need to provide your Google credentials to allow it to access the
\code{Participants} database in Google Drive. You will also be prompted to write
the formr password in the console. Please do not write your password in any
script you will save later.
}
\examples{
my_email <- "gonzalo.garciadecastro@upf.edu"
ml_connect(google_email = my_email)
}
