% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_connect.R
\name{ml_connect}
\alias{ml_connect}
\title{Authenticate in Google and formr}
\usage{
ml_connect(google_email = NULL, formr_password = NULL)
}
\arguments{
\item{google_email}{E-mail used in Google Drive account. If NULL (default), it is assumed to be the same as formr_email.}

\item{formr_password}{Password used in formr account. Do NOT hard-code (include it in a script) this password at any time. We suggest using the keyring R package to store your encrypted password to avoid introducing it every time. Run \code{keyring::key_set(service = "formr", username = "myusername")}, and introduce your password. Your password will be locally store and encrypted. You can retrieve you password running \code{keyring::key_get("formr", "myusername")}. This way, your password will not be hard-coded in your scripts.}
}
\value{
TRUE if Google and formr authentication was successful, FALSE if authentication of any of the two failed.
}
\description{
Authenticate in Google and formr
}
\examples{
my_email <- "gonzalo.garciadecastro@upf.edu"
ml_connect(formr_password = keyring::key_get("multilex", my_email))
}
