% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_logs.R
\name{ml_logs}
\alias{ml_logs}
\title{Generate participant information and progress for each response}
\usage{
ml_logs(
  participants = NULL,
  responses = NULL,
  bilingual_threshold = 5,
  other_threshold = 10
)
}
\arguments{
\item{participants}{Participants data frame, as generated by \code{ml_participants}. If NULL (default), \code{ml_participants} is run.}

\item{responses}{Responses data frame, as generated by \code{ml_responses}. If NULL (default), \code{ml_responses} is run.}

\item{bilingual_threshold}{Minimum degree of exposure to consider a participant as *monolingual*.}
}
\value{
A dataset with participant-level information (one row per response)
}
\description{
Generate participant information and progress for each response
}
\examples{
l <- ml_logs(runs = c("formr-short", "formr-2"))

}
