% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_norms.R
\name{ml_norms}
\alias{ml_norms}
\title{Generate item-level norms for age, sex, language profile and item dominance}
\usage{
ml_norms(
  participants = NULL,
  responses = NULL,
  logs = NULL,
  google_email = NULL,
  norms_item = NULL,
  norms_language = c("Catalan", "Spanish"),
  norms_type = c("understands", "produces"),
  norms_age = c(0, 100),
  norms_lp = c("Bilingual", "Monolingual", "Other"),
  norms_sex = c("Female", "Male"),
  norms_category = NULL,
  conf = 0.95
)
}
\arguments{
\item{participants}{Participants data frame, as generated by \code{ml_participants}. If NULL (default), \code{ml_participants} is run.}

\item{responses}{Responses data frame, as generated by \code{ml_responses}. If NULL (default), \code{ml_responses} is run.}

\item{google_email}{E-mail used in Google Drive account.}

\item{norms_language}{Language to compute vocabulary norms for: "catalan" and/or "spanish}

\item{norms_type}{Vocabulary type to compute norms for: "understands", "produces"}

\item{norms_age}{Age range of participants to compute norms for. It takes a numeric vector of length two (min-max).}

\item{norms_lp}{Language profile of participants to compute norms for: "Bilingual", Monolingual", "Other"}

\item{norms_sex}{Sex of participants to compute norms for: "Female", "Male". Both included by default.}

\item{norms_category}{Semantic/functional category to include items from}

\item{conf}{Confidence level of confidence intervals. Defaults to 0.95.}
}
\value{
A dataset with each participant's comprehensive and/or vocabulary size in each language
}
\description{
Generate item-level norms for age, sex, language profile and item dominance
}
\examples{
vocabulary <- ml_norms(google_email = "user@email.com"))

}
