% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_responses.R
\name{ml_responses}
\alias{ml_responses}
\title{Retrieve and update local and/or remote data from formr}
\usage{
ml_responses(
  participants = NULL,
  runs = c("BL-Long2", "BL-Lockdown"),
  longitudinal = "all",
  update = TRUE
)
}
\arguments{
\item{participants}{Participants data frame, as generated by \code{ml_participants}. If NULL (default), \code{ml_participants} is run.}

\item{runs}{formr runs to update}

\item{longitudinal}{Should longitudinal data be included? If "all" (default), all responses (including repeated measures) are included. If "no", participants with more than one responses to the questionnaire (regardless of the version) are excluded. If "first", only the first response of each participant is included. If "last", only the last response of each participant is included. If "only", only responses with repeated measures are included.}
}
\value{
A data frame containing participant's responses to each item, along with some session-specific metadata.
}
\description{
Retrieve and update local and/or remote data from formr
}
\examples{
d <- ml_responses(runs = c("formr-short", "formr-2"), longitudinal = "last")
}
