% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_vocabulary.R
\name{ml_vocabulary}
\alias{ml_vocabulary}
\title{Generate participant information and progress for each response}
\usage{
ml_vocabulary(
  participants = NULL,
  responses = NULL,
  by = NULL,
  scale = "count"
)
}
\arguments{
\item{participants}{Participants data frame, as generated by \code{ml_participants}. If NULL (default), \code{ml_participants} is run.}

\item{responses}{Responses data frame, as generated by \code{ml_responses}. If NULL (default), \code{ml_responses} is run.}

\item{by}{A character vector that takes the name of the variable(s) to group data into. Vocabulary metrics will be calculated by aggregating responses within the groups that result from the combination of crossing of the variables provided in \code{by}. This variables can refer to item properties (see \code{pool}, e.g., "category") or to participant properties (see \code{ml_logs()}, e.g., "dominance").}

\item{scale}{A character vector that takes the value "count" and/or "prop". If "count" (default) vocabulary metrics are reported as counts (number of words). If "prop", vocabulary metrics are calculated as proportions?}
}
\value{
A dataset with each participant's comprehensive and/or vocabulary size in each language
}
\description{
Generate participant information and progress for each response
}
\examples{
vocabulary <- ml_vocabulary()

}
