% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_wordcloud.R
\name{ml_wordcloud}
\alias{ml_wordcloud}
\title{Retrieve and update local and/or remote data from formr}
\usage{
ml_wordcloud(responses = NULL, participant)
}
\value{
A plot with the wordcloud of one participant across ages
}
\description{
Retrieve and update local and/or remote data from formr
}
\examples{
words <- ml_wordcloud(responses = ml_responses(), participant = "bilexicon_000")
}
